refactoring server queries
