+++ lib/types/parameterized-query.js
@@ -13,9 +13,8 @@ const {QueryFile} = require('../query-file');
 const {assertOptions} = require('assert-options');
 
 const npm = {
-    os: require('os'),
-    utils: require('../utils'),
-    errors: require('../errors')
+    EOL: require('os').EOL,
+    utils: require('../utils')
 };
 
 /**
@@ -91,7 +90,6 @@ class ParameterizedQuery extends ServerFormatting {
 
     constructor(options) {
         super();
-
         if (typeof options === 'string' || options instanceof QueryFile) {
             options = {
                 text: options
@@ -99,32 +97,8 @@ class ParameterizedQuery extends ServerFormatting {
         } else {
             options = assertOptions(options, ['text', 'values', 'binary', 'rowMode']);
         }
-
-        const _i = {
-            options,
-            changed: true,
-            currentError: undefined,
-            PQ: {},
-            setValues(v) {
-                if (Array.isArray(v)) {
-                    if (v.length) {
-                        _i.PQ.values = v;
-                    } else {
-                        delete _i.PQ.values;
-                    }
-                } else {
-                    if (npm.utils.isNull(v)) {
-                        delete _i.PQ.values;
-                    } else {
-                        _i.PQ.values = [v];
-                    }
-                }
-            }
-        };
-
-        _i.setValues(options.values);
-
-        npm.utils.addReadProp(this, '_inner', _i, true);
+        this._inner.options = options;
+        this._inner.setValues(options.values);
     }
 
     /**
@@ -136,17 +110,6 @@ class ParameterizedQuery extends ServerFormatting {
      * Only the basic variables (`$1`, `$2`, etc) can be used in the query, because _Parameterized Queries_
      * are formatted on the server side.
      */
-    get text() {
-        return this._inner.options.text;
-    }
-
-    set text(value) {
-        const _i = this._inner;
-        if (value !== _i.options.text) {
-            _i.options.text = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name ParameterizedQuery#values
@@ -159,13 +122,6 @@ class ParameterizedQuery extends ServerFormatting {
      * - None of the above, means it is a single formatting value, which
      *   is then automatically wrapped into an array
      */
-    get values() {
-        return this._inner.PQ.values;
-    }
-
-    set values(value) {
-        this._inner.setValues(value);
-    }
 
     /**
      * @name ParameterizedQuery#binary
@@ -176,17 +132,6 @@ class ParameterizedQuery extends ServerFormatting {
      *
      * @see {@link http://www.postgresql.org/docs/devel/static/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY Extended Query}
      */
-    get binary() {
-        return this._inner.options.binary;
-    }
-
-    set binary(value) {
-        const _i = this._inner;
-        if (value !== _i.options.binary) {
-            _i.options.binary = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name ParameterizedQuery#rowMode
@@ -196,17 +141,6 @@ class ParameterizedQuery extends ServerFormatting {
      * Changes the way data arrives to the client, with only one value supported by $[pg]:
      *  - `array` will make all data rows arrive as arrays of values. By default, rows arrive as objects.
      */
-    get rowMode() {
-        return this._inner.options.rowMode;
-    }
-
-    set rowMode(value) {
-        const _i = this._inner;
-        if (value !== _i.options.rowMode) {
-            _i.options.rowMode = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name ParameterizedQuery#error
@@ -218,9 +152,6 @@ class ParameterizedQuery extends ServerFormatting {
      *
      * This property is primarily for internal use by the library.
      */
-    get error() {
-        return this._inner.currentError;
-    }
 
     /**
      * @method ParameterizedQuery#parse
@@ -238,11 +169,11 @@ class ParameterizedQuery extends ServerFormatting {
         const qf = options.text instanceof QueryFile ? options.text : null;
 
         if (!_i.changed && !qf) {
-            return _i.PQ;
+            return _i.target;
         }
 
-        const errors = [], values = _i.PQ.values;
-        _i.PQ = {
+        const errors = [], values = _i.target.values;
+        _i.target = {
             text: options.text
         };
         _i.changed = true;
@@ -253,33 +184,33 @@ class ParameterizedQuery extends ServerFormatting {
             if (qf.error) {
                 errors.push(qf.error);
             } else {
-                _i.PQ.text = qf[QueryFile.$query];
+                _i.target.text = qf[QueryFile.$query];
             }
         }
 
-        if (!npm.utils.isText(_i.PQ.text)) {
+        if (!npm.utils.isText(_i.target.text)) {
             errors.push('Property \'text\' must be a non-empty text string.');
         }
 
         if (!npm.utils.isNull(values)) {
-            _i.PQ.values = values;
+            _i.target.values = values;
         }
 
         if (options.binary !== undefined) {
-            _i.PQ.binary = !!options.binary;
+            _i.target.binary = !!options.binary;
         }
 
         if (options.rowMode !== undefined) {
-            _i.PQ.rowMode = options.rowMode;
+            _i.target.rowMode = options.rowMode;
         }
 
         if (errors.length) {
-            return _i.currentError = new ParameterizedQueryError(errors[0], _i.PQ);
+            return _i.currentError = new ParameterizedQueryError(errors[0], _i.target);
         }
 
         _i.changed = false;
 
-        return _i.PQ;
+        return _i.target;
     }
 }
 
@@ -318,7 +249,7 @@ ParameterizedQuery.prototype.toString = function (level) {
         lines.push(gap + 'error: ' + this.error.toString(level + 1));
     }
     lines.push(npm.utils.messageGap(level) + '}');
-    return lines.join(npm.os.EOL);
+    return lines.join(npm.EOL);
 };
 
 module.exports = {ParameterizedQuery};
