+++ lib/types/prepared-statement.js
@@ -13,9 +13,8 @@ const {QueryFile} = require('../query-file');
 const {assertOptions} = require('assert-options');
 
 const npm = {
-    os: require('os'),
-    utils: require('../utils'),
-    errors: require('../errors')
+    EOL: require('os').EOL,
+    utils: require('../utils')
 };
 
 /**
@@ -93,34 +92,9 @@ class PreparedStatement extends ServerFormatting {
 
     constructor(options) {
         super();
-
         options = assertOptions(options, ['name', 'text', 'values', 'binary', 'rowMode', 'rows']);
-
-        const _i = {
-            options,
-            changed: true,
-            currentError: undefined,
-            PS: {},
-            setValues(v) {
-                if (Array.isArray(v)) {
-                    if (v.length) {
-                        _i.PS.values = v;
-                    } else {
-                        delete _i.PS.values;
-                    }
-                } else {
-                    if (npm.utils.isNull(v)) {
-                        delete _i.PS.values;
-                    } else {
-                        _i.PS.values = [v];
-                    }
-                }
-            }
-        };
-
-        _i.setValues(options.values);
-
-        npm.utils.addReadProp(this, '_inner', _i, true);
+        this._inner.options = options;
+        this._inner.setValues(options.values);
     }
 
     /**
@@ -154,17 +128,6 @@ class PreparedStatement extends ServerFormatting {
      * Changing this property for the same {@link PreparedStatement#name name} will have no effect, because queries
      * for Prepared Statements are cached by the server, with {@link PreparedStatement#name name} being the cache key.
      */
-    get text() {
-        return this._inner.options.text;
-    }
-
-    set text(value) {
-        const _i = this._inner;
-        if (value !== _i.options.text) {
-            _i.options.text = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name PreparedStatement#values
@@ -177,13 +140,6 @@ class PreparedStatement extends ServerFormatting {
      * - None of the above, means it is a single formatting value, which
      *   is then automatically wrapped into an array
      */
-    get values() {
-        return this._inner.PS.values;
-    }
-
-    set values(value) {
-        this._inner.setValues(value);
-    }
 
     /**
      * @name PreparedStatement#binary
@@ -194,17 +150,6 @@ class PreparedStatement extends ServerFormatting {
      *
      * @see {@link http://www.postgresql.org/docs/devel/static/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY Extended Query}
      */
-    get binary() {
-        return this._inner.options.binary;
-    }
-
-    set binary(value) {
-        const _i = this._inner;
-        if (value !== _i.options.binary) {
-            _i.options.binary = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name PreparedStatement#rowMode
@@ -214,17 +159,6 @@ class PreparedStatement extends ServerFormatting {
      * Changes the way data arrives to the client, with only one value supported by $[pg]:
      *  - `array` will make all data rows arrive as arrays of values. By default, rows arrive as objects.
      */
-    get rowMode() {
-        return this._inner.options.rowMode;
-    }
-
-    set rowMode(value) {
-        const _i = this._inner;
-        if (value !== _i.options.rowMode) {
-            _i.options.rowMode = value;
-            _i.changed = true;
-        }
-    }
 
     /**
      * @name PreparedStatement#rows
@@ -254,9 +188,6 @@ class PreparedStatement extends ServerFormatting {
      *
      * This property is primarily for internal use by the library.
      */
-    get error() {
-        return this._inner.currentError;
-    }
 
     /**
      * @method PreparedStatement#parse
@@ -275,18 +206,18 @@ class PreparedStatement extends ServerFormatting {
         const qf = options.text instanceof QueryFile ? options.text : null;
 
         if (!_i.changed && !qf) {
-            return _i.PS;
+            return _i.target;
         }
 
-        const errors = [], values = _i.PS.values;
-        _i.PS = {
+        const errors = [], values = _i.target.values;
+        _i.target = {
             name: options.name,
             text: options.text
         };
         _i.changed = true;
         _i.currentError = undefined;
 
-        if (!npm.utils.isText(_i.PS.name)) {
+        if (!npm.utils.isText(_i.target.name)) {
             errors.push('Property \'name\' must be a non-empty text string.');
         }
 
@@ -295,36 +226,36 @@ class PreparedStatement extends ServerFormatting {
             if (qf.error) {
                 errors.push(qf.error);
             } else {
-                _i.PS.text = qf[QueryFile.$query];
+                _i.target.text = qf[QueryFile.$query];
             }
         }
-        if (!npm.utils.isText(_i.PS.text)) {
+        if (!npm.utils.isText(_i.target.text)) {
             errors.push('Property \'text\' must be a non-empty text string.');
         }
 
         if (!npm.utils.isNull(values)) {
-            _i.PS.values = values;
+            _i.target.values = values;
         }
 
         if (options.binary !== undefined) {
-            _i.PS.binary = !!options.binary;
+            _i.target.binary = !!options.binary;
         }
 
         if (options.rowMode !== undefined) {
-            _i.PS.rowMode = options.rowMode;
+            _i.target.rowMode = options.rowMode;
         }
 
         if (options.rows !== undefined) {
-            _i.PS.rows = options.rows;
+            _i.target.rows = options.rows;
         }
 
         if (errors.length) {
-            return _i.currentError = new PreparedStatementError(errors[0], _i.PS);
+            return _i.currentError = new PreparedStatementError(errors[0], _i.target);
         }
 
         _i.changed = false;
 
-        return _i.PS;
+        return _i.target;
     }
 }
 
@@ -367,7 +298,7 @@ PreparedStatement.prototype.toString = function (level) {
         lines.push(gap + 'error: ' + this.error.toString(level + 1));
     }
     lines.push(npm.utils.messageGap(level) + '}');
-    return lines.join(npm.os.EOL);
+    return lines.join(npm.EOL);
 };
 
 module.exports = {PreparedStatement};
